requirements.txt
sqlite-vec
requests
watchdog
mss
pillow
winocr
fastapi
uvicorn[standard]
pyqt6
